+++ test/fg/flagsmith_caches_test.dart
@@ -18,6 +18,11 @@ void main() {
       expect(() => fs.hasCachedFeatureFlag(notImplmentedFeature),
           throwsA(isA<FlagsmithConfigException>()));
     });
+
+    test('When caches not enabled and get cached flag then fail', () async {
+      expect(() => fs.getCachedFeatureFlagValue(notImplmentedFeature),
+          throwsA(isA<FlagsmithConfigException>()));
+    });
   });
 
   group('[Caches] enabled', () {
@@ -33,6 +38,15 @@ void main() {
       await fs.getFeatureFlags();
       final _value = fs.hasCachedFeatureFlag(notImplmentedFeature);
       expect(_value, false);
+    });
+
+    test('When caches enabled then get value', () async {
+      await fs.getFeatureFlags();
+      final _value = fs.hasCachedFeatureFlag(notImplmentedFeature);
+      expect(_value, false);
+
+      final _flagValue = fs.getCachedFeatureFlagValue('min_version');
+      expect(_flagValue, '2.0.0');
     });
 
     test('When feature flag remove then success', () async {